feat(cppclient): provide entry points for C API (#5695)
This PR provides the entry points for the "C API" to the Deephaven C++
library.

Notes for the reviewer: `client_options_interop.[cc,h]` is probably the
easiest to understand so I would recommend starting there. Then
`client_interop.[cc,h]`

In many cases this is just writing "glue code" mechanically just by
looking at the signatures of the methods in the C++ classes.

In a few cases we have made exceptions for simplicity.

For example `TableHandle::Sort` takes a vector of `SortPair`. For the C
API, we *could* take an array of SortPair objects. But I decided it was
simpler to decompose the SortPair object into its three components:
`column_name_`, `direction_`, and `abs_`. So, for the Sort() C API we
just pass in three arrays: of column names, of directions, and of abs,
plus their length. Then the C API doesn't need to define a "SortPair"
object at all.

There is also a bunch of template-heavy code involving changes to
`TableMaker`. The C# code relies on this heavily (to make temporary
tables for use in tests). You may not need to drill into this code,
because it is complicated.

Currently the C++ test code, such as in `select_test.cc`, uses a call to
the templated `CompareTables` method to compare actual tables to
expected tables. In a future PR I will modify all this code to use
`TableMaker` instead. At that time it will probably be more obvious how
`TableMaker` is used.