chore: Upgrade jedi to 0.19.1 (#5693)
Adds support for Python 3.11 and 3.12.

Also, the 0.19.0 changelog notes a massive performance improvement for
`Interpreter` mode which we use by default. It felt a bit snappier to me
while using it locally.

Also adds a new flag to avoid unsafe execution if we want to enable
that. Not sure what the default is based on the changelog.

https://jedi.readthedocs.io/en/latest/docs/changelog.html