build: Remove java classpath conventions (#5724)
This removes our custom configuration hierarchy, historically inherited
from the "fishlib" days.

Follow-up to #5683

Note: unlike #5683, this migration does not aim to preserve the exact
configuration dependencies that were inherited in the
java-classpath-conventions hierarchy. The configuration conventions were
overly broad, and didn't represent the real dependencies that each
project has. As part of this, our dependency on
`org.jetbrains:annotations` changed from an implementation dependency
(inherited through the `fishBase` configuration) to a compileOnly
dependency to better reflect how we actually depend on it. We might
expect that the jetbrains annotations would be removed from our
application runtime classpath entirely, but that is not the case - it
turns out that one of our transitive dependencies has a runtime
dependency on it:

```
org.jetbrains:annotations:13.0                                                                                                                                                                                       
\--- org.jetbrains.kotlin:kotlin-stdlib:1.9.10                                                                                                                                                                       
     +--- io.confluent:kafka-protobuf-provider:7.6.0                                                                                                                                                                 
     |    \--- io.confluent:kafka-protobuf-serializer:7.6.0                                                                                                                                                          
     |         \--- project :extensions-kafka                                                                                                                                                                        
     |              \--- project :server                                                                                                                                                                             
     |                   \--- project :server-jetty                                                                                                                                                                  
     |                        \--- runtimeClasspath
```

In this case, it has the visible effect of "downgrading" our runtime
classpath from what was previously `org.jetbrains:annotations:24.0.0` to
`org.jetbrains:annotations:13.0`. Arguably, this is the more correct
runtime dependency to provide given it matches what the exact version
the transitive library requests.