build: Make it easier to manually run GWT tests from the IDE (#5725)
Summary from the new task's description:

Test wiring to run either unit or integration tests with a manual
browser and an already-running server. This makes it easier to run a
tests repeatedly, either one at a time or as a class/suite, without
paying to start/stop selenium and deephaven each time. The port will
remain constant at 8888 each run to let breakpoints continue to work
across repeated runs.
 
To use this, first start a server on port 10000 with anonymous access
enabled. Then, either select a test in IntelliJ to run using the
manualGwtTest task, or invoke from the command line with info logging
enabled and a specific test selected, e.g.:

./gradlew :web-client-api:manualGwtTest --info --tests
io.deephaven.web.client.api.NullValueTestGwt

Click the URL that is printed out to run the test in your browser, or
refresh an existing browser window.

Co-authored-by: Mike Bender <mikebender@deephaven.io>