fix: correct floating point ChunkHasher (#5778)
This adds documentation and a new hashCode method to the
`io.deephaven.util.compare` package that is consistent with Deephaven
equality. Specifically, this ensures that the floating points values
-0.0 and 0.0 hash to the same value.

Testing was added around aggregation keys, join keys, aggregations,
min/max formulas, and sort results. Of note is that unique and distinct
aggregations currently rely on
`io.deephaven.chunk.WritableChunk#sort()`, which treats -0.0 < 0.0. This
effects the encounter order returned by these aggregations, and could be
seen as inconsistent with our stated goals of treating -0.0 == 0.0. This
does not affect the consistency of the sort operation. It is hard to be
confident that the testing coverage around this issue is fully complete;
that said, this does make us much more consistent in the tested
operations wrt Deephaven equality.

This is a prerequisite for #5605

Fixes #3768