perf: Remove unused naturalJoin key states. (#5770)
The generated hash tables for naturalJoin (and other operations) never
removed a bucket's hash table entry when it became entry. For cases
where a naturalJoin cycles through buckets, the hash table could be
filled with empty values consuming memory without reason.

This changes the generated open addressed hash table for both-sides
incremental naturalJoin to mark entries deleted with a tombstone when
both the left and right side are empty. The occupancy of the hash table
for use in load factor computation includes the number of entries in the
hash table and tombstones. The rehash operation may copy entries into a
new table that is the same size as the original table instead of always
doubling in size, allowing tombstone entries to be reaped.

Closes #5769

---------

Co-authored-by: Ryan Caudy <rcaudy@gmail.com>