fix: document and update WritableChunk and CompactKernel
The WritableChunk documentation and implementation was update to more accurately reflect the desired behavior. A "fix-up" stage for floating point types was introduced. A sortUnsafe method was introduced to skip the "fix-up" stage for callers who do not need the "fix-up".

The CompactKernel documentation and implementation was updated to more accurately reflect the desired behavior.

The add-only min max operators were updated to ignore NaN values.

Fixes #5824