feat(cpp-client): Add GetTidAsString as a utility function (#5843)
Add `GetTidAsString()` as a utility function.

Note that the previous version of this code was `inline`. However, on
the Windows platform, making it inline involves pulling in some windows
includes including `windows.h`. The problem with including `windows.h`
is that it defines some macros that might conflict. In particular it
defines the `min` and `max` macros that definitely conflict.

In order to free ourselves and our Windows customers from seeing the
conflicts that might arise, we isolate the Windows-specific code in a
.cc file rather than including `windows.h` in a .h file.

There are other approaches that would work too, such as defining
`NOMINMAX` or changing all of our calls that look like `min()` and
`max()` to have extra parentheses: `(min)()` and `(max)()`.

However in my opinion, isolating this Windows-specific code in one .cc
file was simplest for now.