fix: Upgrade to gradle
This upgrades from the no longer maintained action `burrunan/gradle-cache-action` to the gradle maintained action `gradle/actions/setup-gradle`. There is a different usage pattern; with `setup-gradle`, it is invoked first during setup as opposed to at the command site. We were also using job-ids with `burrunan/gradle-cache-action` to implicitly control cache prefixes; it is not clear if we actually need to separate caches, and so (for now), the default caching behavior of `gradle/actions/setup-gradle` is used, https://github.com/gradle/actions/blob/v3.5.0/docs/setup-gradle.md#caching-build-state-between-jobs. By default, the action will read from cache, but will only write to the cache for jobs on the `main` branch.

Fixes #5161