feat: detect piped union
This allows vermin to correctly warn us if we are using the pipe union syntax. Useful to prevent issues like #5912 in the future. In addition, self-documenting fstrings will be reporteed. Both of these features are labelled as "unstable"; if vermin ever produces a false positive, we can re-consider the inclusion of these features.

For example, it would have produced the following error:

```
Detecting python files..
Analyzing 247 files using 24 processes..
!2, 3.10     /home/devin/dev/deephaven/deephaven-core/py/server/deephaven/_table_reader.py
  union types as `X | Y` require !2, 3.10

Minimum required versions: 3.10
Incompatible versions:     2
Target versions not met:   3.8
```

Here is the documentation for these feature flags:

```
  [--feature <name>] ...
        Some features are disabled by default due to being unstable:
          fstring-self-doc - [Unstable] Detect self-documenting fstrings. Can in
                             some cases wrongly report fstrings as self-documenting.
          union-types      - [Unstable] Detect union types `X | Y`. Can in some cases
                             wrongly report union types due to having to employ heuristics.
```