feat: log session ID and user-agent
Logs the session ID and user-agent (if set) on session creation. Technically, the user-agent could change with each RPC call, but that is not something that clients typically do. For contexts where logging every session is too spammy, the configuration property `SessionListenerLogger.level` can be used to set the appropriate log level.

Note, the user agent interceptor is not able to be bound to `Set<ServerInterceptor>` as it causes an undefined ordering with respect to `SessionServiceInterceptor`; we need to ensure that the user agent interceptor is applied before the session service interceptor.

Server component of #5704