fix!: Fix a TimeRange off-by-one bug in nanosecond calculation (#5648)
Fix a TimeRange off-by-one bug in nanosecond calculation.

Resolves #5646 

BREAKING CHANGE: The length of a business day was being incorrectly
computed by 1 ns. To keep the business day length the same for existing
calendars, the close of a business period has been changed from
inclusive to exclusive. This default behavior can be overridden by using
`<includeClose>true</includeClose>` when specifying time ranges. This
will include the closing time for the range and will result in a
business period that is 1 ns longer. Business calendars using the legacy
format must be upgraded to the current format to use
`<includeClose>true</includeClose>`.