feat(csharp/ExcelAddIn): Make connections in the background (#6011)
The "Test Connection" and "Set Connection" button could delay your
progress because they were serialized on a single thread.

This can get annoying if you type in the wrong credentials, as computer
might take several seconds to figure that out. Even if you correct your
credentials in the meantime, you would have to wait for the original
connection request to resolve.

This PR changes the code so that it does that work on a background
thread.

Doing so also exposed a bug in my organization of Observables, so they
have been reorganized a little bit.

In particular I was trying to be clever by having one object observe two
different things and the logic wasn't quite right.