fix(py-client-test): Read PluginClients In Order; fix shared_ticket Ref (#6155)
This is related to investigation in #5996. Swap the order of these tests to reduce
the probability of the py-grpc client from claiming the rpc failed.

One test was not testing the proper shared ticket; adjusted for correctness to test intended feature.

ObjectService change is to adopt the proper ExportObject dependency pattern. Otherwise 
correctness requires analysis/knowledge that the list of objects are all brand new server side 
exports.