chore: Update lychee ignore file (#6300)
I was under the impression that to do substring matching, the tool would
need asterisks via regex in the file. That didn't work. I found an
example of another repo using this tool and looked at their lychee
ignore file to find no asterisks.

This past weekend's runs failed on links that contained localhost and
docs.deephaven.io. This should fix that.

There is a legit broken link in the run that will be fixed as part of
this PR as well.