fix: update S3ParquetTestBase testReadWriteUsingProfile
This fixes an nightly error that #6305 introduced where `S3ParquetLocalStackTest.testReadWriteUsingProfile` was not erroring out with the use of an "invalid" region name. While we might want to have some sort of test like this, it should be standalone and not tied to the usage of profiles. As such, the corresponding code was removed.

Ultimately, the behavior was traced down to a difference in behavior between localstack 3.6 and localstack 3.7; the region name logic must have changed in some way. We could do a deep dive into the commits to figure out why the behavior changed, or file an issue, if we care to pursue this further.

https://github.com/localstack/localstack/releases/tag/v3.7.0