fix: update S3ParquetTestBase testReadWriteUsingProfile (#6310)
This fixes an nightly error that #6305 introduced where
`S3ParquetLocalStackTest.testReadWriteUsingProfile` was not erroring out
with the use of an invalid profile or credentials data (as far as the
earlier images were concerned, localstack was failing based on the
region and MinIO was failing based on the keys). While we might want to
have some sort of test like this, it should be standalone and not tied
to the usage of profiles. The invalid keys based test was separated out
and only enabled for MinIO. An "invalid" region does not trigger this
error for MinIO nor Localstack.

Ultimately, the behavior was traced down to a difference in behavior
between localstack 3.6 and localstack 3.7; the logic must have changed
in some way. We could do a deep dive into the commits to figure out why
the behavior changed, or file an issue, if we care to pursue this
further.

https://github.com/localstack/localstack/releases/tag/v3.7.0