fix: Rename directory containing proto files (#6379)
In order to make directory structures usable for downstream projects
seeking to add new services or messages that depend on our own proto
files, ensure that we name/package consistently for all languages.

This may be a breaking change for some downstream consumers, but these
changes should prevent needing to make future changes or add additional
build steps. Two places this may impact applications:
* The Python client proto package has been renamed from
`pydeephaven.proto` to `deephaven_core.proto`. The old import will
continue to work in this release, but will warn the first time it is
referenced.
* The deephaven-proto-backplane-grpc.jar still contains the .proto
files, but they are located in the deephaven_core/proto directory
instead of at the root of the jar.

Fixes #6376