feat(csharp/ExcelAddIn): Add a version number and display it in the GUI (#6602)
We are establishing a version numbering scheme for the Excel AddIn.

This PR sets the AddIn to be version 0.8.0, displays that version in the
dialog boxes, and we also note the versions of Community Core and
Enterprise Core+ that it was built from in the file VERSIONS.md.

We also make relevant changes to the ExcelAddInInstaller that also
surface the version number for the installer.