fix: Improve JS API handling of formatting data (#6664)
Cells with formatted rows would error when read in two specific ways:
either if the subscription/snapshot call was made before the table to be
read was resolved, or if the format data was read directly from the
TableData rather than from the row or the column.

Additionally fixes a bug where an unresolved table with a viewport being
created on it cannot be snapshotted.

Fixes DH-18634
Fixes DH-18784