fix: Close ManifestReader in IcebergBaseLayout.findKeys
This fixes a ManifestReader leak in IcebergBaseLayout, which was not being closed. Additionally, this refactors the findKeys implementation to use Iterable patterns instead of Stream patterns, making the code less lamba / wrap heavy. A couple of unused IcebergUtils methods were removed, and one is marked for test use only since none of the actual Iceberg implementation code depends on it anymore.