feat(cpp-client): Eliminate the member variable time_nano_scale_factor_ [part 1/8 of groupby] (#6787)
The current version of this code calculates the needed "scale factor" at
construction time and caches it in a member variable.

This is not necessary because the column source knows its underlying
Arrow type and can calculate the needed scale factor on demand.

This change removes that member variable.

There is no user-visible behavioral change.