build: Fix two oversights in our published jars
First, we leave bytecode in published code that references
util-immutables annotations, but we don't make those annotations
available. This fixes that - but does not make the annotations a
requirement for downstream projects. Any project which wants to depend
on the annotations themselves can do so.

Second, Gradle does not include apt-generated sources in its published
source jars. This patch adds those generated types to the source jar.