feat(cpp-client): Move method to make it available to others [pyticking 4/7] (#6838)
This PR moves a piece of functionality into a more central location
(namely, ContainerUtil::Inflate) so it can be used by other code. In a
later PR, we will call this method from the CythonSupport code.

Note: this PR is a clone of
https://github.com/deephaven/deephaven-core/pull/6829. Github automation
does not seem to want to allow me to merge that PR (possibly because I
closed and reopened it), so I'm trying the same code here on a "fresh"
PR.