feat(cpp-client): Add more support in CythonSupport for nested lists [pyticking 5/7] (#6830)
This PR is dependent on
https://github.com/deephaven/deephaven-core/pull/6829 and should not be
reviewed until that PR is merged

This PR allows C++ to do some of the "heavy lifting" for working with
Arrow nested lists.

Because, broadly speaking, the interface between Python and the `dhcore`
library is over `ColumnSource`, this PR adds more methods to
`CythonSupport` to work with nested lists.

One of the methods, `SlicesToColumnSource`, takes leaf data like
`[a,b,c,d,e,f,null,g]` and length data like `[3,null,0,4]`, both passed
in as `ColumnSources`, and creates a new `ColumnSource<ContainerBase>`
containing `[[a,b,c],null,[],[e,f,null,g]]`

Another method, `ContainerToColumnSource`, takes a `ContainerBase<T>`
and wraps it as a `ColumnSource<T>`. This is useful because the Cython
code already knows how to convert ColumnSources into python arrays, so
it's useful to have a `ColumnSource` representation of a `Container` so
we don't have to write a bunch more Python code.