feat(cpp-client): Remove redundant ElementTypeId support from Schema [pyticking 7/7] (#6832)
This PR is dependent on
https://github.com/deephaven/deephaven-core/pull/6831 and should not be
reviewed until that PR is merged

This PR removes a backward-compatible accessor in our `Schema` class
that is no longer needed now that the Cython code has been updated.

Also I happened to notice that `Schema` had its copy constructors and
assignment operators enabled. They should be disabled (a) on general
principles and (b) copies won't even work because of the `string_view`
trick I'm doing.