fix: DH-19570: UnionSourceManager Should Not Inspect Stale Slot Location Entries (#6932)
We were not bounds checking the initial array lookup. When losing
partitions we don't zero out or change any values in the
slot-lookup-array that are beyond the total number of slots. However the
number of valid slots may decrease from prev to curr (or vice versa) and
a lookup past the end of an array could appear as if the slot mapping
was right/valid, but really it's just looking at stale values and making
bad choices.

The fix is to bounds check. The unit test is targeted and failed before
the fix.