docs: Remove requirements from Python readme (#6931)
The requirements section is duplicative of our core docs. It makes no
sense to document in two places, because that will make it more
difficult to keep up with.

There is still useful information in this page, so we shouldn't remove
it altogether. This just links to the core docs for requirements.