feat: add AWS SessionCredentials to S3 module (#6941)
The current version of the Deephaven S3 module does not support
temporary AWS credentials via
[SessionCredentials](https://docs.aws.amazon.com/AmazonS3/latest/API/API_SessionCredentials.html).
The Amazon SDK supports this, so this PR adds the necessary wrapping in
the Java and Python DH layers to expose SessionCredentials to the user.
I did not write any tests for the Python layer as it appears that the
"basic" credentials, which also requires specific creds, is untested. I
assume this means it is not trivial to write such tests without real
credentials of some kind.