fix: DH-19656: Fix Windows build of Deephaven Core (#6945)
This PR fixes three problems that were preventing Community Core from
building successfully on Windows. Additionally, it removes some
debugging code that had some silly messages.

Briefly,
1. The way we do Windows builds, we rely on cmake to do the necessary
dllexport/dllimport stuff for shared libraries. However that support is
limited to functions; it does not work for global variables. The
workaround we use is that when we have a global variable that we want to
provide from one library and reference in another, we represent that
global variable as a simple accessor function. In a previous PR I forgot
about this limitation and introduced a bunch of global variables.
2. There were a few instances of Unicode character literals that Visual
Studio wasn't parsing correctly, probably because it was assuming the
source file had an iso-latin-1 encoding rather than UTF-8. This was
fixed by using \uxxxx escape codes.
3. There was an expression in feedtimes.cc that calculated the
difference between two time_points. The Windows compiler couldn't find a
suitable `operator-`. It's possible this is because
`std::chrono::high_resolution_clock` has a different type on Windows and
Linux. By specifying types to `ToTimePoint` explicitly we are able to
make this work.
4. Removed the infelicitous method `zambonidump` from
cython_support_test.cc