fix: NPE in UnionSourceManager When Constituent Fails (#6956)
This is the point release PR for #6953 on 0.39.x.

A community user ran into this when trying to merge several
barrage-subscribed tables. They were all failing, causing an NPE as the
iterator advanced past the end of the list and then attempts to
dereference a null instead of capturing the error locally before
advancing the iterator.