fix: DH-18563, DH-19804: Migrate shadow plugin to com.gradleup.shadow (#6980)
The johnrengelman shadow plugin has been migrated to the GradleUp
organization (see https://github.com/GradleUp/shadow/issues/908).

In addition, the custom POM xml building has been replaced with setting
a custom artifact name (as documented
https://gradleup.com/shadow/publishing/#publish-the-shadowed-jar-with-custom-artifact-name).

This is a separation of some work that would otherwise make DH-19690
more complicated.

This also fixes DH-19804 by ensuring the grpc-bom is correctly referred
to in the `deephaven-mtls-authentication-provider` pom (in the
`dependencyManagement` section).