feat: Improve JobScheduler interface with cleanup callback
This improves the JobScheduler interface to present a cleanup callback, allowing callers to more easily deal with releasing resources without forcing the caller to maintain their own atomic boolean. The contract is that exactly one of onCompleteCleanup or onError will be invoked. (This is the same sort of pattern that gRPC StreamObserver presents.)

This will be used by #6958

This PR also fixes an improper RowSet usage in `RegionedColumnSourceManager#pushdownFilter` similar to #6959