fix: DH-19338: Ensure we have an ExecutionContext while handling HierarchicalTableViewSubscription snapshots. (#6999)
Cherry-pick of https://github.com/deephaven/deephaven-core/pull/6984

This consists of three actual changes:
1. The HierarchicalTableView subscriptions would call snapshot, etc.
from their update propagation job. This resulted in sorts that had no
context. In the Core+ case, this meant we could not correctly log
performance in the face of sorted rollups. This is the root of this
entire PR.
2. The FunctionGeneratedTableFactory had a similar problem when
refreshing a table.
3. I added a flag to PerformanceEntry that will fail when the auth
context is not set. It defaults to false, to avoid breaking Core users
unnecessarily. In Enterprise, we are breaking already; so this is an
appropriate setting. I also added it to the tests, to find other cases
where we were broken (like the FunctionGeneratedTableFactory).

The rest of this is just changing the tests to include a context.

I think #1 and #2 are a point release candidate, but not the rest of the
fixes/changes.

---------

Co-authored-by: Ryan Caudy <rcaudy@gmail.com>