feat: DH-19045: FilterBarriers and FilterRespectsBarriers (#6885)
Introduces `withBarrier` and `withRespectsBarrier` to the Filter API. 

There is support to prevent reordering of a filter that respects a
barrier around a filter that declares that barrier in the following
locations:
- `PartitionAwareSourceTable.PartitionAwareTableReference`
- `PartitionAwareSourceTable#where`
- `QueryTable#initializeAndPrioritizeFilters`
- `AbstractFilterExecution.StatelessFilter#compareTo`
- `DeferredViewTable`

This change introduces a `WhereFilter.Visitor` targeted for at
`WhereFilterImpl`s that wrap one or more `WhereFilter`s.
This change also rewrites how `QueryTable#where` integrates with
constant-offset-column-array access; now the same shift is shared across
all filters even when they are embedded within complex filter
structures.