feat: DH-19978: Add assertBlink, assertAddOnly and assertAppendOnly to Table. (#7037)
The existing TableAssertions.assertAppendOnly has been removed, as it
does not actually verify that the table is append-only (it was checking
for add-only) and is worse in terms of listeners.

To make tables fire their assertions, even without a listener; BaseTable
notifyListeners is slightly reorganized.

The comment on removeBlink has more exposition (which in part addresses
DH-19918).

---------

Co-authored-by: Ryan Caudy <rcaudy@gmail.com>