fix: DH-20018: Update Python Arrow to Table Converter Type Mapping (#7055)
This allows more types to be used via the Arrow to Table Converter via
python. These types had barrage support added earlier this year.

At this time, the result table is often going to be non-displayable to
the web-ui. This is because the web-ui does not drop the original
uploaded schema before fetching data, and jsapi is unable to parse some of
these column types in their original arrow column format.