feat: DH-20077: Simplify client builds and prepare for CI automation (#7065)
Note to reviewers: the intention here is to make a single consolidated
build script (and instructions) for all the Windows clients: C++,
Python, eventually R, ...

One question is where this script should live. It's currently in
cpp-client which is the wrong place (too specific). But we don't have a
directory for "all clients". Thoughts?