refactor: DH-20135: migrate javapoet
square’s javapoet library is deprecated https://github.com/square/javapoet?tab=readme-ov-file#deprecated

The recommended replacement is https://github.com/palantir/javapoet. It seems to be a clean fork of the original with bug fixes (probably new features as well), with ongoing maintenance and releases. A quick audit of the repo shows it is using gradle instead of maven.