refactor: DH-20135: migrate javapoet (#7082)
square’s javapoet library is deprecated
https://github.com/square/javapoet?tab=readme-ov-file#deprecated
    
The recommended replacement is https://github.com/palantir/javapoet. It
seems to be a clean fork of the original with bug fixes, new features,
ongoing maintenance, and releases. A quick audit of the repo shows it is
using gradle instead of maven.

A couple of code changes were necessary. 1) Instead of referencing field
members directly, we must now reference the proper method. 2) Extraneous
arguments can no longer be passed to code generation methods.

This also fixes authorization codegen documentation to correctly
reference the directory `deephaven_core` (as opposed to `deephaven`);
this was an oversight from #6379.