fix(cpp-client): DH-20266: Improve C++ Windows build times by providing the /MP flag (#7138)
The flag permits compiling concurrently with multiple cores. We have
wanted to do this before but we never landed on the correct set of
flags.

This PR:
1. Adds the `/MP` flag to the dhclient, dhcore, and tests projects
2. Removes `/p:CL_MPCount=16` from `build-windows-clients.bat`. Removing
this flag allows the build system to choose the "right" number of cores
to use rather than setting to be a fixed 16.
3. Keeps the "m" flag but changes the character controlling it from `-`
to `/`. Both work but we change it to `/` for consistency.

(BTW: the m flag controls the concurrency of building multiple PROJECTS
at the same time, and we may not be ready for that kind of concurrency)